+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-There are no changes yet.
+### Changed
+
+- CSS specificity for `.plot:nth-of-type(4n+i)` and `.plot .element:nth-of-type(4n+i)` is lowered.
 
 
 ## [1.0.0-alpha.1] - 2022-03-05